# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Cellar Service
  description: HTTP service for managing your wine cellar
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /api-product/product/{id}:
    get:
      tags:
      - product
      summary: product product
      description: 商品信息
      operationId: product#product
      parameters:
      - name: id
        in: path
        description: auctionId
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ProductProductResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/ProductProductNotFoundResponseBody'
            required:
            - message
            - id
      schemes:
      - http
...
definitions:
...
  MediaResponseBody:
    title: MediaResponseBody
    type: object
    properties:
      content:
        type: integer
        description: 视频链接
        example: 1
        format: int64
      media_id:
        type: integer
        description: Id
        example: 169686
        format: int64
      media_type:
        type: integer
        description: 媒体类型：0-图片，1-视频，2-未知
        example: 0
        format: int64
      media_url:
        type: string
        description: 图片URL
        example: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
      sequence:
        type: integer
        description: 上传顺序，默认为0
        example: 1
        format: int64
    example:
      content: 1
      media_id: 169686
      media_type: 0
      media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
      sequence: 1
  ProductProductNotFoundResponseBody:
    title: ProductProductNotFoundResponseBody
    type: object
    properties:
      id:
        type: string
        description: ID of missing bottle
        example: Et sunt necessitatibus enim laboriosam officia.
      message:
        type: string
        description: Message of error
        example: bottle 1 not found
    description: Bottle not found
    example:
      id: Quo eum optio voluptates excepturi consectetur.
      message: bottle 1 not found
    required:
    - message
    - id
  ProductProductResponseBody:
    title: 'Mediatype identifier: product; view=default'
    type: object
    properties:
      art_no:
        type: string
        description: 编号
        example: ZS1908038
      carouselList:
        type: array
        items:
          $ref: '#/definitions/MediaResponseBody'
        example:
        - content: 1
          media_id: 169686
          media_type: 0
          media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
          sequence: 1
        - content: 1
          media_id: 169686
          media_type: 0
          media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
          sequence: 1
      category_id:
        type: integer
        example: 4482537499051224988
        format: int64
      categoryName:
        type: string
        example: 18K金镶嵌翡翠
      cert_code:
        type: integer
        example: 2533094998625962217
        format: int64
      cert_type:
        type: integer
        example: 5430983705901641484
        format: int64
      colour:
        type: string
        example: Dignissimos delectus quis aut culpa sapiente.
      colour_id:
        type: integer
        example: 843583157850833356
        format: int64
      created_at:
        type: integer
        example: 9148560457983445096
        format: int64
      crowd:
        type: string
        example: Voluptates necessitatibus beatae et possimus.
      current_price:
        type: integer
        example: 5462332762198272151
        format: int64
      detailPics:
        type: array
        items:
          $ref: '#/definitions/MediaResponseBody'
        example:
        - content: 1
          media_id: 169686
          media_type: 0
          media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
          sequence: 1
        - content: 1
          media_id: 169686
          media_type: 0
          media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
          sequence: 1
        - content: 1
          media_id: 169686
          media_type: 0
          media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
          sequence: 1
        - content: 1
          media_id: 169686
          media_type: 0
          media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
          sequence: 1
      end_time:
        type: integer
        example: 4054326239755873016
        format: int64
      extAttrMap:
        type: integer
        example: 606180961827252239
        format: int64
      id:
        type: integer
        description: 商品Id
        example: 15272
        format: int64
      level:
        type: integer
        example: 3248837449528108781
        format: int64
      level_id:
        type: integer
        example: 1946610086700734394
        format: int64
      mkt_price:
        type: integer
        example: 756138538825062831
        format: int64
      name:
        type: string
        example: Suscipit accusantium.
      operator_id:
        type: integer
        example: 3442971926677727333
        format: int64
      recommendations:
        type: string
        example: Ipsam distinctio illum.
      scenario:
        type: string
        example: Facilis quod.
      scenarioList:
        type: array
        items:
          type: string
          example: "1"
        example:
        - "1"
        - "1"
      size:
        type: string
        example: Ut deserunt deleniti.
      status:
        type: integer
        example: 4988229028296355710
        format: int64
      style:
        type: string
        example: Aperiam voluptatem officia iusto.
      style_id:
        type: string
        example: Fuga rerum minima.
      summary:
        type: string
        example: Nam nam praesentium quis.
      template_id:
        type: integer
        example: 905726755904667550
        format: int64
      updated_at:
        type: integer
        example: 5763374062725037415
        format: int64
      version:
        type: integer
        example: 3275925694035348337
        format: int64
      weight:
        type: string
        example: Blanditiis rerum expedita amet.
    description: ProductResponseBody result type (default view)
    example:
      art_no: ZS1908038
      carouselList:
      - content: 1
        media_id: 169686
        media_type: 0
        media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
        sequence: 1
      - content: 1
        media_id: 169686
        media_type: 0
        media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
        sequence: 1
      category_id: 4326852687865262773
      categoryName: 18K金镶嵌翡翠
      cert_code: 7644296537778396651
      cert_type: 6321302060432734401
      colour: Rem voluptatem assumenda delectus corrupti commodi.
      colour_id: 9179107296571209048
      created_at: 6638066124886776828
      crowd: Consequuntur voluptatibus quia et.
      current_price: 1604456867846254238
      detailPics:
      - content: 1
        media_id: 169686
        media_type: 0
        media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
        sequence: 1
      - content: 1
        media_id: 169686
        media_type: 0
        media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
        sequence: 1
      - content: 1
        media_id: 169686
        media_type: 0
        media_url: https://img.cft.upmi.com.cn/o_1d3o3vcmu10723kgta21qb31v25a.jpg
        sequence: 1
      end_time: 8302124091010216481
      extAttrMap: 3546869669647755401
      id: 15272
      level: 6823684965968876884
      level_id: 8721319520708291317
      mkt_price: 4157482281085871412
      name: Facilis officiis quod.
      operator_id: 3630304078323438114
      recommendations: Quia tenetur dolores totam.
      scenario: Voluptatem assumenda ullam.
      scenarioList:
      - "1"
      - "1"
      - "1"
      size: Laborum in qui ut sint et dignissimos.
      status: 8006147338144456856
      style: Voluptatem officia tempora aut aut possimus alias.
      style_id: Totam perferendis ullam voluptates placeat quod assumenda.
      summary: Excepturi voluptatibus tempora est fugit velit molestiae.
      template_id: 2023082694273874914
      updated_at: 4119404029315103872
      version: 2134172924211955258
      weight: Odit facilis fugiat accusamus ut inventore quod.
...